feat(persistence): A1 migrate auth stores into persistence/{sqlite,postgres}/
- Move SqliteSessionStore/SqliteLockoutStore/SqliteRefreshStore +
  Postgres variants from api/auth/*_store.py into
  persistence/{sqlite,postgres}/{session,lockout,refresh}_repo.py with
  rename to *Repository.
- Extract RefreshRecord to api/auth/refresh_record.py so the protocol
  can reference it without cycling through the deleted refresh_store.
- Expose backend.sessions + backend.refresh_tokens as properties on
  PersistenceBackend. LockoutRepository is built via
  backend.build_lockouts(auth_config) since it requires the operator's
  threshold/window/duration policy.
- Delete _build_session_store / _build_lockout_store /
  _build_refresh_store dispatcher triplet from api/lifecycle.py;
  startup reads backend properties directly.
- Delete api/auth/session_store.py, lockout_store.py, refresh_store.py.
- Update api/state.py, auth/__init__.py, auth/controller.py, and
  affected test files to import from new persistence locations.
- Issue #1457.
