+++ src/synthorg/api/auth/__init__.py
@@ -4,11 +4,6 @@ from synthorg.api.auth.config import AuthConfig
 from synthorg.api.auth.models import ApiKey, AuthenticatedUser, AuthMethod, User
 from synthorg.api.auth.service import AuthService
 from synthorg.api.auth.session import Session
-from synthorg.api.auth.session_store import (
-    PostgresSessionStore,
-    SessionStore,
-    SqliteSessionStore,
-)
 from synthorg.api.auth.ticket_store import TicketLimitExceededError, WsTicketStore
 
 __all__ = [
@@ -17,10 +12,7 @@ __all__ = [
     "AuthMethod",
     "AuthService",
     "AuthenticatedUser",
-    "PostgresSessionStore",
     "Session",
-    "SessionStore",
-    "SqliteSessionStore",
     "TicketLimitExceededError",
     "User",
     "WsTicketStore",
