+++ tests/unit/api/auth/test_session_store.py
@@ -12,8 +12,11 @@ import aiosqlite
 import pytest
 
 from synthorg.api.auth.session import Session
-from synthorg.api.auth.session_store import SessionStore, SqliteSessionStore
 from synthorg.api.guards import HumanRole
+from synthorg.persistence.auth_protocol import SessionRepository as SessionStore
+from synthorg.persistence.sqlite.session_repo import (
+    SQLiteSessionRepository as SqliteSessionStore,
+)
 
 pytestmark = pytest.mark.unit
